+++ premium/web-frontend/modules/baserow_premium/components/crud_table/CrudTableColumn.js
@@ -7,9 +7,11 @@
  * @param cellComponent The cell component class used to render a cell of this column,
  * crudTable will listen to 'edit-row' (expected to emit a single updated row
  * object) and
- * 'delete-row' (expected to emit a  events from this cell and
+ * 'delete-row' (expected to emit an updated row) events from this cell and
  * update the rows accordingly. CrudTable will pass two props to this component, the
- * column which is this object and the specific row returned by the service.
+ * column which is this object and the specific row returned by the service. Finally
+ * any listeners defined on the CrudTable itself will be passed through to all
+ * instances of this cellComponent.
  * @param {string} minWidth A valid min parameter to the css minmax(HERE,) function
  * used to construct CrudTable's css grid column templates.
  * @param {string} maxWidth A valid max parameter to the css minmax(,HERE) function to