+++ premium/web-frontend/modules/baserow_premium/services/userAdmin.js
@@ -1,5 +1,11 @@
 export default (client) => {
   return {
+    /**
+     * @param {number} page The page number to fetch.
+     * @param {{searchQuery}} searchQuery A search term to filter users by username.
+     * @param {[{direction:string, key:string}]} sorts An ordered list of sorts to
+     * apply to the users.
+     */
     fetchPage(page, searchQuery, sorts) {
       const params = { page }
       if (searchQuery) {