+++ web-frontend/modules/core/components/Paginator.vue
@@ -12,7 +12,7 @@
         <i class="fas fa-caret-left"></i>
       </a>
       <input
-        v-model="textInputPage"
+        v-model.number="textInputPage"
         class="input paginator__page-input"
         type="number"
         @keypress.enter="changePage(textInputPage)"
@@ -32,17 +32,22 @@
 </template>
 <script>
 /**
+ * Shows pagination buttons and the current page. Emits a `change-page` event when the
+ * user attempts to change the page. If this event is successfully completed it is up
+ * the parent component to update the provided page prop to show the new current page.
  */
 export default {
   name: 'Paginator',
   props: {
     /**
+     * The total number of pages available.
      */
     totalPages: {
       required: true,
       validator: (prop) => typeof prop === 'number' || prop === null,
     },
     /**
+     * The currently selected page.
      */
     page: {
       required: true,
@@ -54,13 +59,22 @@ export default {
       textInputPage: 1,
     }
   },
+  watch: {
+    page(newPage) {
+      this.textInputPage = newPage
+    },
+  },
   methods: {
+    invalidNewPage(newPage) {
+      return (
+        typeof newPage !== 'number' ||
+        (this.totalPages !== null &&
+          this.totalPages !== 0 &&
+          (newPage > this.totalPages || newPage < 1))
+      )
+    },
     changePage(newPage) {
-      if (
-        this.totalPages !== null &&
-        this.totalPages !== 0 &&
-        (newPage > this.totalPages || newPage < 1)
-      ) {
+      if (this.invalidNewPage(newPage)) {
         this.textInputPage = this.page
         return
       }