+++ web-frontend/modules/database/components/view/ViewSearchContext.vue
@@ -66,6 +66,7 @@ export default {
       lastSearch: '',
       hideRowsNotMatchingSearch: true,
       lastHide: true,
+      loading: false,
     }
   },
   methods: {